rzero: updates for building sdcard image
We now use the vanilla 5.10 tree and patch in the radxa zero patches on top.
Also included are 2 hacks, one to remove a shutdown callback that fixes rebooting the devices, as well as one that gets rid of an error early in boot, and finally a patch to recognize the processor revision.

Update the firmware to the newest 43456 that I could find (comes rpi-distro/firmware-nonfree, aka the firmware from the RaspberryPi 400)

Also add in a script that generates a uInitrd after creating the initrd, so that u-boot will load it properly.